docs: fix development setup instructions in README.md
setup-db.py requires a third argument as it also allows to update a
database so 0 needs to be explicitly given for full initialization.
(TODO: maybe default to 0 if not provided?)

It also requires the database uri in order to properly initalize the
database. The secret key is not necessary for the operation but needed
for the application initalization.